[fix] : 문제&시계열데이터 관계수정 1:N -> N:N

+++ src/main/java/com/example/radialapi/survey/dto/TimeDataLevelDto.java
@@ -9,9 +9,9 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class TimeDataLevelDto {
     private LocalDateTime time;
-    private int dataLevel;
+    private double dataLevel;
 
-    public TimeDataLevelDto(LocalDateTime time, int dataLevel) {
+    public TimeDataLevelDto(LocalDateTime time, double dataLevel) {
         this.time = time;
         this.dataLevel = dataLevel;
     }
